// src/core/domain/integrationConfigs/enums/stringTimeInterval.enum.ts
export enum STRING_TIME_INTERVAL {
    LAST_12_HOURS = '-12h',
    LAST_24_HOURS = '-24h',
    LAST_48_HOURS = '-48h',
    LAST_72_HOURS = '-72h',
    LAST_7_DAYS = '-7d',
    LAST_14_DAYS = '-14d',
    LAST_20_DAYS = '-20d',
    LAST_21_DAYS = '-21d',
    LAST_1_MONTH = '-1M',
    LAST_2_MONTHS = '-2M',
    LAST_3_MONTHS = '-3M',
}